requirements.txt
git+https://git@github.com/ndustrialio/ndustrialio-python.git@SPIKE/assetsII#egg=ndustrialio-python
tabulate
dateutils
